Restructure project into MVC packages
