+++ src/main/java/sim/app/Order.java
@@ -1,4 +1,4 @@
-package sim.coffee;
+package sim.app;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
