+++ src/main/java/sim/view/Observer.java
@@ -1,8 +1,8 @@
-package sim.coffee;
+package sim.view;
 
 //Reference: PatternsExampleCode
 public interface Observer {
-	
+
 	public void update();
 
 }
